Moved ElementCascadingDeleter to a different file, lint fixes
